fix: Food 엔티티 ID 타입을 Long에서 Integer로 변경 (DB SERIAL 타입 매칭)

+++ backend-spring/src/main/java/com/NurtiAgent/Onboard/food/repository/FoodRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface FoodRepository extends JpaRepository<Food, Long> {
+public interface FoodRepository extends JpaRepository<Food, Integer> {
 
     /**
      * 음식 검색 (그룹화 + 평균)
